artifacts/api-server-py/pyproject.toml
[project]
name = "predictive-api"
version = "0.0.0"
description = "FastAPI backend for the predictive engine"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "asyncpg>=0.30.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "neo4j>=5.26.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
